Add documentation to the hub.js, handle adding mis typed hubs better
For adding and updating hub I've decided to return the response.d-
ata if it succeeds but a boolean false if it does not.  The reaso-
ning for this is to enable easier handling of the error message,no
matter what gets passed down from the API as an error, it still g-
ets handled in a uniform and neat way.