Datatable with promise integration
This is proving to be a huuuuge bitch when it comes to refreshing
the data WHEN you click save.  This is not an easy feat, I've tri-
ed a hundred different methods, you'll see that refresh and new d-
ata buttons work (which was also a huge pain in the ass) but I've
got it now.  You'll also notice that I need to have the html page
directly declare ng-controller, I hate this.  But for it to work
it needs to happen, it seems there is no way around it. Walking
away for now.