Refactor and document new functions. @shirazamin read commit msg
So everything here is done the proper way for the data tables,
since we are fetching promises datatables will not always render
if the data isn't fetched quick enough.  As a result we need to
use the promise method of inserting data into the table.
With that comes much more things to be done, but I have it all
done!  You should copy this for the users page! It's all
commented out and pretty as it could be, so it should be easy to
 repeat :)