Hacky fix to view hub from the Datatable on mobile
I am not too sure how to override this and really nothing comes up
in google.  So what happens is on mobile, the first item is a
clickable, it will make the table expand to show the rest of the
details. I can't find any documentation or how to override this
anywhere. It's incredibly unintuitive and I literally discovered
it by accident.  Another big problem with this, is that the button
inside the dropdown that gets formed won't work, I am not sure why

To fix this I just switched the order of status and the view
details button.  It'll work for now, there are bigger fish to fry.
fixes #23