Set default hub in the profile page
Doesn't work, I have no idea why.  it is correctly setting the
profile.default_hub_id to the selected one, it is correctly
sending that into the auth.updateAccount, it is correctly changing
the users name.  But when it comes to setting the default hub it
refuses.  I have tried it passing in as a string, and as of now
it is casting the value.