Charts filled with data
This is so far from ideal or what I even want, it'll probably get
dubbed.

The problem: I can't think of any other way to do it.  I want only
2 charts, that have multiple series for each sensor and then you
can easily see all the data of each sensor of it's kind side by
side.  The problem resides in there being ykeys, that are kind of
hard coded in.  It's a bit of a bitch, I can't think of a way to
dynamically change this, I've messed around with several ways but
I am getting nothing.